fix: inconsistent doucment naming
